Adds a new way of marking the version
The version was set by hardcoding it in the code. Now it's taken care by
the build system (CMake).

First we check if there is a version file (included in release tarballs),
then try git-describe, then use the default. A VERSION_FILE is generated
so external package generators, e.g., NSIS.
A re-configure is forced on each git commit so that you can trust the
values of the variables in your build system.

The GetGitRevisionDescription functions written by Ryan Pavlik.
(https://github.com/rpavlik/cmake-modules)

Signed-off-by: Luís Pereira <luis.artur.pereira@gmail.com>