Adds Preferences/Settings framework classes
It's a very limited framework. Qt should provide an preferences/settings
framework (IMHO).

PreferencesDialog provides an icon list configuration dialog box. It
supports Ok, Apply and Cancel buttons.

PreferencesPageInterface is and interface for the preferences pages managed
by PreferencesDialog.

SettingsInterface is the interface for the keeper of each page options.

Signed-off-by: Luís Pereira <luis.artur.pereira@gmail.com>