Make the AboutDialog close button save geometry too
The close button emits the rejected() signal, who was connected to the
QWidget::reject() slot. But.... the QWidget::reject() slot doesn't emit an
QCloseEvent. But the QWidget::close() slot indeed emits it.
Connecting the rejected() signal to the close() slot will create an
QCloseEvent that will be catched by AboutDialog::closeEvent() which calls
writeSettings().

Signed-off-by: Luís Pereira <luis.artur.pereira@gmail.com>